Ensure CopyII, CopyRAI match 'copy.img' alias
This makes sure that we can copy any type of image
